+++ R/defaults.R
@@ -97,6 +97,7 @@ opts_chunk_attr = local({
   opts$fig.ext = as.list(unique(auto_exts))
   opts$external = opts$sanitize = NULL  # hide these two rare options
   opts$fig.process = 'function'
+  opts$fig.asp = 'numeric'
   opts$R.options = 'list'
   opts$cache.comments = 'logical'
   opts